# docs/swagger.yaml
definitions:
  model.Todo:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /todos:
    get:
      consumes:
      - application/json
      description: Возвращает список всех задач
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Todo'
            type: array
        "500":
          description: Ошибка при получении задач
          schema:
            type: string
      summary: Получить все задачи
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Создает задачу с переданными данными
      parameters:
      - description: Задача
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/model.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Todo'
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "500":
          description: Не удалось создать задачу
          schema:
            type: string
      summary: Создать новую задачу
      tags:
      - todos
  /todos/{id}:
    get:
      consumes:
      - application/json
      description: Удаляет задачу с указанным ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Todo'
        "404":
          description: Задача не найдена
          schema:
            type: string
        "500":
          description: Ошибка при получении задачи
          schema:
            type: string
      summary: Удалить задачу по ID
      tags:
      - todos
swagger: "2.0"